requirement.txt
flask
selenium
beautifulsoup4
requests
prometheus_client
psycopg2-binary
